+++ bpm-engine-impl/src/main/java/io/takari/bpm/form/FormService.java
@@ -2,6 +2,7 @@ package io.takari.bpm.form;
 
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.model.form.FormDefinition;
+import io.takari.bpm.model.form.FormField;
 
 import java.util.Map;
 import java.util.UUID;
@@ -41,4 +42,14 @@ public interface FormService {
      * @throws ExecutionException
      */
     FormSubmitResult submit(UUID formInstanceId, Map<String, Object> data) throws ExecutionException;
+
+    /**
+     * Converts the supplied data into a form field.
+     * @param m input data. Typically a map with a single entry: {@code key -> value},
+     *          where {@code key} is the name of the field and {@value} are the field's options.
+     * @return a form field.
+     */
+    default FormField toFormField(Map<String, Object> m) {
+        throw new IllegalStateException("Not implemented");
+    }
 }