+++ bpm-engine-impl/src/main/java/io/takari/bpm/form/FormTaskHandler.java
@@ -3,23 +3,22 @@ package io.takari.bpm.form;
 import io.takari.bpm.ProcessDefinitionUtils;
 import io.takari.bpm.actions.CreateEventAction;
 import io.takari.bpm.api.ExecutionContext;
-import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.api.ExecutionContextFactory;
+import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.api.Variables;
 import io.takari.bpm.model.ProcessDefinition;
 import io.takari.bpm.model.UserTask;
 import io.takari.bpm.model.UserTask.Extension;
 import io.takari.bpm.model.form.FormDefinition;
 import io.takari.bpm.model.form.FormExtension;
+import io.takari.bpm.model.form.FormField;
 import io.takari.bpm.state.ProcessInstance;
 import io.takari.bpm.state.StateHelper;
 import io.takari.bpm.task.UserTaskHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Collection;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 public class FormTaskHandler implements UserTaskHandler {
 
@@ -51,7 +50,9 @@ public class FormTaskHandler implements UserTaskHandler {
         log.debug("handle ['{}', '{}', '{}'] -> found form extension: {}",
                 state.getBusinessKey(), definitionId, elementId, x);
 
-        FormDefinition fd = formDefinitionProvider.getById(x.getFormId());
+        Map<String, Object> options = getOptions(x, state.getVariables());
+
+        FormDefinition fd = getOrCreateFormDefinition(x, options);
         if (fd == null) {
             throw new ExecutionException("Form definition not found: " + x.getFormId());
         }
@@ -59,7 +60,6 @@ public class FormTaskHandler implements UserTaskHandler {
         String pk = state.getBusinessKey();
         UUID fId = UUID.randomUUID();
         String eventName = UUID.randomUUID().toString();
-        Map<String, Object> options = getOptions(x, state.getVariables());
         Map<String, Object> env = state.getVariables().asMap();
 
         formService.create(pk, fId, eventName, fd, options, env);
@@ -67,9 +67,41 @@ public class FormTaskHandler implements UserTaskHandler {
         return StateHelper.push(state, new CreateEventAction(definitionId, elementId, eventName, null, null, null, null));
     }
 
+    @SuppressWarnings("unchecked")
     private Map<String, Object> getOptions(FormExtension x, Variables vars) {
         ExecutionContext ctx = contextFactory.create(vars);
         return (Map<String, Object>) ctx.interpolate(x.getOptions());
+    }
+
+    private FormDefinition getOrCreateFormDefinition(FormExtension x, Map<String, Object> options) throws ExecutionException {
+        Object v = options != null ? options.get("fields") : null;
+        if (v == null) {
+            return formDefinitionProvider.getById(x.getFormId());
+        }
+
+        if (!(v instanceof List)) {
+            throw new IllegalArgumentException("Expected a list of fields in form '" + x.getFormId() + "', got: " + v.getClass());
+        }
+
+        return new FormDefinition(x.getFormId(), coerceToFormFields((List<?>) v));
+    }
+
+    @SuppressWarnings("unchecked")
+    private List<FormField> coerceToFormFields(List<?> input) {
+        List<FormField> result = new ArrayList<>(input.size());
+
+        for (Object v : input) {
+            if (v instanceof FormField) {
+                result.add((FormField) v);
+            } else if (v instanceof Map) {
+                FormField f = formService.toFormField((Map<String, Object>) v);
+                result.add(f);
+            } else {
+                throw new IllegalArgumentException("Expected either a FormField instance or a map of values, got: " + v);
+            }
+        }
+
+        return result;
     }
 
     private static FormExtension findFormExtension(Collection<Extension> extensions) {